+++ tests/Node/TextTest.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 use PHPHtmlParser\Dom;
 use PHPHtmlParser\Dom\Node\TextNode;
+use PHPHtmlParser\Options;
 use PHPUnit\Framework\TestCase;
 use stringEncode\Encode;