+++ CHANGELOG.md
@@ -5,8 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 3.1.1
+
+### Changed
+- Fixed issue with numbers in comments.
+- Updated minimume php version to correct version.
+- Comment tags are now self-closing when cleanup input is set to false.
+
 ## 3.1.0
 
+### Changed
 - Updated to include Tidelift subscription option.
 - Removed php-coverall.
 - Removed Guzzle 6 Adapter.